+++ toys-core/test/lookup-cases/preloads/ns-2/.preload.rb
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+$toys_preload_ns2 = :hi